+++ tests/benchmarks/test_rabbitmq_cli.py
@@ -44,7 +44,7 @@ def latency():
 remoulade.declare_actors([throughput, fib, latency])
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 @pytest.mark.benchmark(group="rabbitmq-100k-throughput")
 def test_rabbitmq_process_100k_messages_with_cli(benchmark, info_logging, start_cli):
     # Given that I've loaded 100k messages into RabbitMQ
@@ -64,7 +64,7 @@ def test_rabbitmq_process_100k_messages_with_cli(benchmark, info_logging, start_
     benchmark.pedantic(broker.join, args=(throughput.queue_name,), setup=setup)
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 @pytest.mark.benchmark(group="rabbitmq-10k-fib")
 def test_rabbitmq_process_10k_fib_with_cli(benchmark, info_logging, start_cli):
     # Given that I've loaded 10k messages into RabbitMQ
@@ -84,7 +84,7 @@ def test_rabbitmq_process_10k_fib_with_cli(benchmark, info_logging, start_cli):
     benchmark.pedantic(broker.join, args=(fib.queue_name,), setup=setup)
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 @pytest.mark.benchmark(group="rabbitmq-1k-latency")
 def test_rabbitmq_process_1k_latency_with_cli(benchmark, info_logging, start_cli):
     # Given that I've loaded 1k messages into RabbitMQ