+++ tests/test_scheduler.py
@@ -11,7 +11,7 @@ from remoulade.scheduler import ScheduledJob
 # - There are no /tmp/schedul* files
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 def test_simple_interval_scheduler(start_scheduler, start_cli):
     from tests.scheduler_configs.simple_1 import broker, write_loaded_at
 
@@ -58,7 +58,7 @@ def test_simple_interval_scheduler(start_scheduler, start_cli):
     assert slow_task.last_queued == last_queued
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 def test_multiple_schedulers(start_scheduler, start_cli):
     from tests.scheduler_configs.simple_1 import broker, write_loaded_at
 
@@ -66,9 +66,7 @@ def test_multiple_schedulers(start_scheduler, start_cli):
     for _ in range(5):
         start_scheduler("tests.scheduler_configs.simple_slow")
 
-    start_cli(
-        "tests.scheduler_configs.simple_slow", extra_args=["--threads", "1"]
-    )
+    start_cli("tests.scheduler_configs.simple_slow", extra_args=["--threads", "1"])
 
     time.sleep(5)
 
@@ -81,14 +79,13 @@ def test_multiple_schedulers(start_scheduler, start_cli):
     assert len(text) == 1
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 @pytest.mark.parametrize("suffix", ["", "_tz"])
 def test_scheduler_daily_time(start_scheduler, start_cli, suffix):
     start_scheduler("tests.scheduler_configs.daily{}".format(suffix))
 
     start_cli(
-        "tests.scheduler_configs.daily{}".format(suffix),
-        extra_args=["--threads", "1"],
+        "tests.scheduler_configs.daily{}".format(suffix), extra_args=["--threads", "1"],
     )
 
     time.sleep(3)
@@ -117,13 +114,11 @@ def test_scheduler_daily_time(start_scheduler, start_cli, suffix):
     assert len(text) == 1
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 def test_scheduler_new_daily_time(start_scheduler, start_cli):
     start_scheduler("tests.scheduler_configs.daily")
 
-    start_cli(
-        "tests.scheduler_configs.daily_new", extra_args=["--threads", "1"]
-    )
+    start_cli("tests.scheduler_configs.daily_new", extra_args=["--threads", "1"])
 
     time.sleep(10)
 
@@ -133,13 +128,11 @@ def test_scheduler_new_daily_time(start_scheduler, start_cli):
             ...
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 def test_scheduler_wrong_weekday(start_scheduler, start_cli):
     start_scheduler("tests.scheduler_configs.weekday_1")
 
-    start_cli(
-        "tests.scheduler_configs.weekday_1", extra_args=["--threads", "1"]
-    )
+    start_cli("tests.scheduler_configs.weekday_1", extra_args=["--threads", "1"])
 
     time.sleep(3)
 
@@ -149,13 +142,11 @@ def test_scheduler_wrong_weekday(start_scheduler, start_cli):
             ...
 
 
-@pytest.mark.skipif(os.getenv("CI") == "1", reason="test skipped on CI")
+@pytest.mark.skipif(os.getenv("CI") == "true", reason="test skipped on CI")
 def test_scheduler_right_weekday(start_scheduler, start_cli):
     start_scheduler("tests.scheduler_configs.weekday_2")
 
-    start_cli(
-        "tests.scheduler_configs.weekday_2", extra_args=["--threads", "1"]
-    )
+    start_cli("tests.scheduler_configs.weekday_2", extra_args=["--threads", "1"])
 
     time.sleep(3)